+++ user-rest-service/domain/userdomain/signup_user_model.go
@@ -41,42 +41,3 @@ func (u *SignUpUser) Email() vo.Email {
 func (u *SignUpUser) Password() vo.Password {
 	return u.password
 }
-
-type LoginUser struct {
-	userID   UserID
-	name     Name
-	email    vo.Email
-	password vo.Password
-}
-
-func NewLoginUser(email vo.Email, password vo.Password) *LoginUser {
-	return &LoginUser{
-		email:    email,
-		password: password,
-	}
-}
-
-func NewLoginUserFromDataSource(userID UserID, name Name, email vo.Email, hashPassword vo.Password) *LoginUser {
-	return &LoginUser{
-		userID:   userID,
-		name:     name,
-		email:    email,
-		password: hashPassword,
-	}
-}
-
-func (u *LoginUser) UserID() UserID {
-	return u.userID
-}
-
-func (u *LoginUser) Name() Name {
-	return u.name
-}
-
-func (u *LoginUser) Email() vo.Email {
-	return u.email
-}
-
-func (u *LoginUser) Password() vo.Password {
-	return u.password
-}
